+++ server/src/main/scala/shared.scala
@@ -92,6 +92,7 @@ object proto {
   case class SetClaimsRequest(uid: String, claims: Map[String, String])
   case class GetLoginMode(email: Email, tenant: Option[TenantCode])
   case class EmitToken(email: Email, magic: Magic, days: Option[Int]) derives JsonCodec
+  case class SwapTokenRequest(token: String, magic: Magic) derives JsonCodec
 
   given JsonDecoder[VerifyGoogleTokenRequest] = DeriveJsonDecoder.gen
   given JsonDecoder[VerifyMorbidTokenRequest] = DeriveJsonDecoder.gen